mainwindow: mark evt parameter as unused
As a practice, I don't leave parameters unnamed because someday someone
may actually want to use it later.  It is quicker for everybody if the
function parameters are clearly defined and the unused parameters
warnings cast away, than a developer wasting time pouring over the
documentation and/or code to work out what the name of the parameter
should be.